Exported apps in src/ to demos/ in root
emphasizes purpose of engine
