+++ demos/demo_sprites/demo_sprites_map.rb
