+++ demos/fractals/fractal_map.rb
