+++ demos/fractals/fractal_meta_data.rb
