+++ demos/game_of_life/game_of_life_map.rb
