+++ demos/pingpong/ping_pong_map.rb
