+++ demos/pingpong/ping_pong_meta_data.rb
