+++ demos/snake/segment.rb
