+++ demos/snake/snake_map.rb
