+++ demos/snake/snake_meta_data.rb
@@ -1,6 +1,7 @@
 require 'game_meta_data'
 require_relative 'snake_map'
-require_relative '../tetris/tetris_achievement_system'
+require 'tetris/tetris_achievement_system'
+
 class SnakeMetaData
   extend GameMetaData
