+++ demos/sokoban/sokoban_map.rb
