+++ demos/sokoban/sokoban_meta_data.rb
@@ -1,7 +1,8 @@
-require_relative '../../game_meta_data'
-require_relative 'demo_sprites_map'
+require 'game_meta_data'
+require_relative 'sokoban_map'
+require 'tetris/tetris_achievement_system'
 
-class DemoSpritesMetaData
+class SokobanMetaData
   extend GameMetaData
 
   def self.theme_list
@@ -21,7 +22,7 @@ class DemoSpritesMetaData
   end
 
   def self.game_map
-    DemoSpritesMap
+    SokobanMap
   end
 
   def self.canvas
@@ -30,13 +31,12 @@ class DemoSpritesMetaData
 
   def self.render_attributes
     {
-        :cell_size => 15,
-        :width_pixels => 40,
-        :height_pixels => 20,
-        :max_width => 600,
-        :max_height => 340,
-        :tics_per_second => 15,
-        :show_grid => false
+        :cell_size => 25,
+        :width_pixels => 20,
+        :height_pixels => 13,
+        :max_width => 500,
+        :max_height => 345,
+        :tics_per_second => 0
     }
   end
 
@@ -46,7 +46,7 @@ class DemoSpritesMetaData
 
   def self.allowed_controls
     {
-      :keyboard => [W_KEY, D_RELEASED, A_RELEASED, A_PRESSED, D_PRESSED],
+      :keyboard => [D_KEY, A_KEY, S_KEY, W_KEY],
       :mouse => []
     }
   end
