+++ demos/tetris/collision_checker.rb
