+++ demos/tetris/shape.rb
