+++ demos/tetris/shapes/l_piece.rb
