+++ demos/tetris/shapes/reverse_l_piece.rb
