+++ demos/tetris/shapes/reverse_squiggly.rb
