+++ demos/tetris/shapes/square.rb
