+++ demos/tetris/shapes/squiggly.rb
